new solution, better performance
Instead of checking to see if numbers are pandigital, it is far more
efficient to generate all permutations of "123", "1234", "12345", etc,
which by definition are all pandigital numbers.  From there, sort by
natural order and check for prime beginning at the end of the list.  The
first prime is the solution.